+++ packages/nimble-components/src/table-column/base/index.ts
@@ -1,9 +1,11 @@
 import {
+    attr,
     ElementStyles,
     observable,
     ViewTemplate
 } from '@microsoft/fast-element';
 import { FoundationElement } from '@microsoft/fast-foundation';
+import { uniqueId } from '@microsoft/fast-web-utilities';
 import type {
     TableCellRecord,
     TableCellState,
@@ -17,6 +19,9 @@ export abstract class TableColumn<
     TCellRecord extends TableCellRecord = TableCellRecord,
     TColumnConfig = unknown
 > extends FoundationElement {
+    @attr({ attribute: 'column-id' })
+    public columnId?: string;
+
     /**
      * @internal
      * Used by the Table in order to give a column a specific pixel width.
@@ -71,6 +76,12 @@ export abstract class TableColumn<
      * This array is parallel with the field names returned from `getDataRecordFieldNames()`.
      */
     public abstract readonly cellRecordFieldNames: readonly TableFieldName[];
+
+    public override connectedCallback(): void {
+        super.connectedCallback();
+
+        this.setAttribute('slot', uniqueId('table-column-slot'));
+    }
 
     /**
      * This method returns the relevant, static configuration a column requires its cellTemplate