+++ packages/nimble-components/src/table/index.ts
@@ -1,4 +1,10 @@
-import { attr, observable, volatile } from '@microsoft/fast-element';
+import {
+    attr,
+    Observable,
+    observable,
+    Notifier,
+    volatile
+} from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import {
     ColumnDef as TanStackColumnDef,
@@ -9,7 +15,7 @@ import {
     getCoreRowModel as tanStackGetCoreRowModel,
     TableOptionsResolved as TanStackTableOptionsResolved
 } from '@tanstack/table-core';
-import type { TableColumn } from '../table-column/base';
+import { TableColumn } from '../table-column/base';
 import { TableValidator } from './models/table-validator';
 import { styles } from './styles';
 import { template } from './template';
@@ -38,8 +44,17 @@ export class Table<
     @observable
     public tableData: TableRowState<TData>[] = [];
 
+    /**
+     * @internal
+     */
     @observable
-    public readonly columns: TableColumn[] = [];
+    public columns: TableColumn[] = [];
+
+    /**
+     * @internal
+     */
+    @observable
+    public readonly childItems: Element[] = [];
 
     /**
      * @internal
@@ -64,6 +79,7 @@ export class Table<
     private readonly table: TanStackTable<TData>;
     private options: TanStackTableOptionsResolved<TData>;
     private readonly tableValidator = new TableValidator();
+    private columnNotifiers: Notifier[] = [];
 
     public constructor() {
         super();
@@ -85,7 +101,42 @@ export class Table<
         this.setTableData(newData);
     }
 
-    public idFieldNameChanged(
+    public override connectedCallback(): void {
+        super.connectedCallback();
+        this.virtualizer.connectedCallback();
+        this.validateAndObserveColumns();
+    }
+
+    public override disconnectedCallback(): void {
+        super.disconnectedCallback();
+        this.virtualizer.disconnectedCallback();
+        this.removeColumnObservers();
+    }
+
+    public checkValidity(): boolean {
+        return this.tableValidator.isValid();
+    }
+
+    /**
+     * @internal
+     *
+     * The event handler that is called when a notifier detects a change. Notifiers are added
+     * to each column, so `source` is expected to be an instance of `TableColumn`, and `args`
+     * is the string name of the property that changed on that column.
+     */
+    public handleChange(source: unknown, args: unknown): void {
+        if (source instanceof TableColumn) {
+            if (args === 'columnId') {
+                this.validateColumnIds();
+            }
+        }
+    }
+
+    protected childItemsChanged(): void {
+        void this.updateColumnsFromChildItems();
+    }
+
+    protected idFieldNameChanged(
         _prev: string | undefined,
         _next: string | undefined
     ): void {
@@ -94,17 +145,51 @@ export class Table<
         this.setTableData(this.table.options.data);
     }
 
-    public override connectedCallback(): void {
-        super.connectedCallback();
-        this.virtualizer.connectedCallback();
+    protected columnsChanged(
+        _prev: TableColumn[] | undefined,
+        _next: TableColumn[]
+    ): void {
+        if (!this.$fastController.isConnected) {
+            return;
+        }
+
+        this.validateAndObserveColumns();
     }
 
-    public override disconnectedCallback(): void {
-        this.virtualizer.disconnectedCallback();
+    private removeColumnObservers(): void {
+        this.columnNotifiers.forEach(notifier => {
+            notifier.unsubscribe(this);
+        });
+        this.columnNotifiers = [];
     }
 
-    public checkValidity(): boolean {
-        return this.tableValidator.isValid();
+    private validateAndObserveColumns(): void {
+        this.removeColumnObservers();
+
+        for (const column of this.columns) {
+            const notifier = Observable.getNotifier(column);
+            notifier.subscribe(this);
+            this.columnNotifiers.push(notifier);
+        }
+
+        this.validateColumnIds();
+    }
+
+    private validateColumnIds(): void {
+        this.tableValidator.validateColumnIds(
+            this.columns.map(x => x.columnId)
+        );
+        this.canRenderRows = this.checkValidity();
+    }
+
+    private async updateColumnsFromChildItems(): Promise<void> {
+        const definedElements = this.childItems.map(async item => (item.matches(':not(:defined)')
+            ? customElements.whenDefined(item.localName)
+            : Promise.resolve()));
+        await Promise.all(definedElements);
+        this.columns = this.childItems.filter(
+            (x): x is TableColumn => x instanceof TableColumn
+        );
     }
 
     /**