+++ packages/nimble-components/src/table/styles.ts
@@ -73,6 +73,7 @@ export const styles = css`
         width: 100%;
         height: 100%;
         position: absolute;
+        pointer-events: none;
     }
 
     .row:hover::before {
@@ -88,6 +89,7 @@ export const styles = css`
                 height: 100%;
                 position: absolute;
                 background: ${fillHoverColor};
+                pointer-events: none;
             }
 
             .row::before {