+++ packages/nimble-components/src/table/tests/table.pageobject.ts
@@ -1,4 +1,5 @@
 import type { Table } from '..';
+import type { TableHeader } from '../components/header';
 import type { TableRecord } from '../types';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 
@@ -16,8 +17,8 @@ export class TablePageObject<T extends TableRecord> {
         return headers.length;
     }
 
-    public getRenderedHeaderContent(columnIndex: number): string {
-        const headers = this.tableElement.shadowRoot!.querySelectorAll(
+    public getHeaderContent(columnIndex: number): Node | undefined {
+        const headers = this.tableElement.shadowRoot!.querySelectorAll<TableHeader>(
             'nimble-table-header'
         )!;
         if (columnIndex >= headers.length) {
@@ -26,7 +27,7 @@ export class TablePageObject<T extends TableRecord> {
             );
         }
 
-        return headers.item(columnIndex).textContent?.trim() ?? '';
+        return this.getHeaderContentElement(headers[columnIndex]!);
     }
 
     public getRenderedRowCount(): number {
@@ -156,4 +157,30 @@ export class TablePageObject<T extends TableRecord> {
         scrollElement.scrollTop = scrollElement.scrollHeight;
         await waitForUpdatesAsync();
     }
+
+    private getHeaderContentElement(
+        element: HTMLElement | HTMLSlotElement
+    ): Node | undefined {
+        const nodeChildren = this.isSlotElement(element)
+            ? element.assignedNodes()
+            : element.shadowRoot?.childNodes;
+        if (!nodeChildren) {
+            return undefined;
+        }
+
+        const slotElement = Array.from(nodeChildren)?.find<HTMLSlotElement>(
+            this.isSlotElement
+        );
+        if (slotElement) {
+            return this.getHeaderContentElement(slotElement);
+        }
+
+        return nodeChildren[0]; // header content should be first item in final slot element
+    }
+
+    private readonly isSlotElement = (
+        element: Node | undefined
+    ): element is HTMLSlotElement => {
+        return element?.nodeName === 'SLOT' ?? false;
+    };
 }