+++ packages/nimble-components/src/wafer-map/styles.ts
@@ -10,7 +10,7 @@ export const styles = css`
 
     .wafer-map-container {
         width: 100%;
-        padding-bottom: 100%;
+        height: 100%;
         position: relative;
         display: inline-block;
         justify-content: center;
@@ -43,16 +43,7 @@ export const styles = css`
         transform: rotate(90deg);
     }
 
-    .zoom-container {
-        width: 100%;
-        height: 100%;
-        position: absolute;
-    }
-
     .circle-base {
-        width: 100%;
-        height: 100%;
-        position: absolute;
         fill: white;
     }
 
@@ -63,17 +54,8 @@ export const styles = css`
         stroke: ${borderColor};
     }
 
-    .wafer-map-area {
-        position: absolute;
-        justify-content: center;
-        align-items: center;
-        width: 100%;
-        height: 100%;
-    }
-
     .wafer-map-canvas {
         display: inline-block;
-        width: 100%;
-        height: 100%;
+        position: absolute;
     }
 `;